Create D&DNext.html
First draft, fully tested. The check boxes for skills are just denote that a character is proficient and the player will need to fill in the appropriate bonus modifier.

The initiative button works with Track Initiative script available in the Roll20 API Library.

All buttons include the modifiers input on the sheets, prompt for targets where applicable, and prompt for miscellaneous modifiers when applicable.