Turn Order for EotE
Added the ability to roll for PC and NPC initiative slots
- !eed npcinit ...  (rolls for NPCs)
- !eed pcinit ... (rolls for PCs)

The results of the roll will look like "X:Y" in the turn order. Where X
is the number of success and Y is the number of advantage.  It will
automatically sort them based on the rules in the CRB.