Create Gumshoe_NBA_Sheet.html
I ran out of time (have 3.5 year old plus work's kicking my butt) before I could figure out how to do stuff I wanted to do, but just typing in stat names is a huge part of the battle for skill-heavy systems. Still, this may be useless to people.

Here are ideas I had:

Currently there are no rolls. Ideally, a roll would include a point spend and deduct from the pool, which seems complicated to code in HTML.

The idea is that you fill in the "max" and, where applicable, the "current" and the rest is handled by maths.

A goal was to have a second tab that was the "setup" tab, where you fill in max values, and the front page was to be current values and spends.

It would also be nice to have lines for languages, contacts, etc. that corresponded to the stat score.