New tabbed character sheet (v3)
New tabbed character sheet, with improved readability, over three tabs :
Main (stats, attacks, skills, special abilities), Numenera & Equipment,
Advancements & Background.
The attributes are the same (so are the macros and API functions), so
the campaign can be updated with the new sheet (HTML and CSS) and keep
existing character(s).