Improvements and Bug Fixes
Corrected the Skills die types to remove d12+* and made it more in line
with the way Attributes are handled with two fields.

Additionally, added all of the Core Skills to the character sheet, and
a configuration section to show which skills should be shown and which
should not be.