Configuration Implemented
Fully implemented the configuration system to allow a user to choose
which background to display, which logo to display and to select from
various setting specific fields.  The overhaul included design layouts
to enhance readability on darker backgrounds and the removal of various
tables (replaced them with DIVs for enhanced functionality in regards
to different fields and UI options.