Additional Setting Support and CSS Cleanup
Added support for cyberpunk settings by adding Strain and Augmentations
as well as Street Cred (for Interface Zero 2.0). Added the IZ2.0 Logo
as an option. Included a couple new Backgrounds. Reorganized the CSS so
that it’s more logical its groupings. Added a Skills Modifier to the
Wild Card skill area and a new heading for Skill Rank.