Updated roll templates
Added Nibrodooh's attack template. modified it slightly so that his template and mine are more closely styled. renamed his template "pathfinder" to "pf_attack" any attack macros that used the old name should be updated.

dropped the fancy header image I was using and just went with white on black.
added rounded corners and a drop shadow to the templates.