Witch Hunter Character Sheet Updates
> Fixed the Skill row alignment; made tables out of everything so now
it should be bullet-proof.
> Added an Advancement tracker inside the Survivor Point section.
Decided against creating a whole new tab since that feature is giving
me headaches on Firefox. Test it out. Because its a repeating field, I
can't make it calculate the whole SP cost total, but it should tell you
the SP cost based on the type of improvement.
> Added "vs Target Discipline" to Charm, Command, Empathy and Perform
rolls. I thought this would be a good reminder (for me as much as
anyone) but it wouldn't require you to target a token on the screen.
It's just a prompt to remind us what Defense affects those skill rolls.
> In the Skills/Talents section, I've added an expanding field called
GM Rolls. This is identical to the customized dice pools at the top of
the section, but the results are only shared with the player and the
GM. I've preconfigured these dice pools for Charm, Notice and Stealth
rolls, but you can customize them however you need. While I prefer
having these as token macros, I can see where players who are leaning
heavily on their characters sheets would find this a helpful option.
> After some confusion, the first row on the weapons table (Combat Tab)
has been revised to allow changes. It defaults to an unarmed attack,
but can be changed by the player.
> Added a max value for the Injury attribute. This should enable the
red status bar on your token when set for Injury.
> Added a "GM Specifics" section. After wrestling with conflicts with
the way I create and maintain monsters and NPCs and the design of the
character sheet, I settled on this as a solution.
> While working on a new set of general purpose macros, I stumbled
across a big headache in the way the sheet calculates armor on the
backend. That has been revised and shouldn't pose an issue anymore. It
shouldn't affect anything you guys see directly, but it should make
sure the weapon damage macros keep working.
> Per request, I've tweaked the Survivor Points section. You now add
new SP to the Unspent SP blank, while Total SP calculates the sum of
Unspent and Spent SP.