Attempt to fix div.sheet-madness bug
Per [this
thread](https://app.roll20.net/forum/post/1691256/css-selector-not-being-processed#post-1691256),
the first selector is not being processed. Because this sheet makes use
of the unicode character Heavy Multiplication X (0x2716), and Roll20's
CSS sanitation process throws out the entire thing if a backslash is
detected, the file must be saved as Unicode. This means the file begins
with the byte-order mark 0xFEFF.

This commit is assuming that the BOM is being inserted into the style
element right next to the ".charsheet" (or right next to the "div") at
the start, which obviously does not match any element on the page, as
there is neither an element `<0xFEFF class="charsheet">` nor an element
`<0xFEFFdiv class="sheet-madness">`. As a result, a newline has been
inserted between the BOM and "div" in the hope that the problem will not
persist.