removed level modifier from attack roll
The rules are worded in a rather confusing manner in m20 purest essence
on page 2.  The level advancement states the changes you make to your
hp, skills, and attack rolls, but fails to mention that this is an
already calculated result.  thus, the added level modifier was adding
WAY more bonsues to attacks than was intended.  Easy misunderstanding of
the rules to make, with the way it's worded.