added "old school" logo and some small sheet adjustments.
Added a logo, changed sex to gender as per roll20 request(I did not change the attribute name "attrib_Sex" so as not to break anyone's macros that pull that attribute), changed the multi-class label on the sheet to "Multi-Class 2" and "Multi-Class 3" accordingly.  Rearranged a few of the attribute labels on the sheet to better organize sheet info.  Added some section titles.