Update HTML and CSS
Changed Gender Field from a Select Field to a Text Field.
Added a “None” to the AC Select List in the Settings Tab.
Added a “Tie Breaker Select in the Settings Tab.
The Initiative roll was changed to add the bonus value of
@{initiative}*@{init-tie}. Multipling the initiative bonus by either 0
or 0.01 depending on the selected value and adding it to the initiative
roll.
Added a Magical Per Plus Die Size to the Weapons in the Inventory
Section.
Added Roll Template information under the Powers in the Powers Tab.
Changed Emote text in the Settings tab to work with Roll Templates.
Added Roll Templates to the character sheet.
Changed the Character Sheet Roll Buttons to use the new Roll Templates
Changed Version number to 2 and Added a Link to the Wiki in the Footer
of the Character Sheet