Create Wilson-Homebrew
This is a basic, flexible prototype sheet for basic roleplaying, regardless of what game system is being used. As I game using a homebrewed system, it works very well across multiple game systems with just a bit of tweeking - we have a homebrewed fantasy system, a version altered for a dystopian Pokemon game, and have used it for modern Call of Cthulhu and Cthulhu by torchlight.