Cypher System v1.2
1) It's now possible to change the style/look of the sheet to the ones
of the generic Cypher System, Numenéra or The Strange.
2) By popular demand, the stat/skill roll template has been unified and
only rolls one d20 dice (for the 3D dice users out there), whatever is
the "Difficulty" roll parameter from the "Action" section. If the
difficulty parameter has been set to 0, the beaten difficulty is no
longer calculated (due to Roll20 limitations), but the raw d20 dice is
shown, as the special natural rolls (1 or 17+), and the steps modifiying
the difficulty are displayed (and summed).
3) New "Artifact" repeating section with integrated Depletion roll.