Implement looping traits for attributes
Code developed at http://jsfiddle.net/Lithl/jovd7L7s/

This implementation of Storyteller System traits is superior to the one
previously used on this sheet and developed for the 2e sheet in that it
permits higher values without taking up additional space. On the other
hand, it also requires using images to display the values.