Update Palladium-Rifts.html
Changed name of "attr_IQ_max" to  "attr_IQmax".

Justification for changing the attribute name:
Some bug in the formula replacement component of Roll20 doesn't evaluate the version with an underscore. This value is trivial to lookup in Palladium; and getting it working saves a lot of effort maintaining the character as it levels up.