Updates to both Html and CSS files
CSS support for a couple of new collapsible sections as well as
additional annotations.

HTML:  add a couple of new collapsible sections, update titles for
repeating sections, update grapple, melee attack bonus, and ranged
attack bonus to have a notes section and a macro-edit section so they
match the style of the rest of the sheet (every roll button should have
a field next to it to edit the macro now).  More work on sheet worker
calculations.