Replaced repeating sections with fixed weapon slots for NPCs
Repeating sections didn't work with macros as macros needed numbered
attributes. So I replaced the repeating sections with a fixed weapon
list for NPCs. So they are identical to the PC sheets now. There is
place for 4 weapons which should be more than enough.

Furthermore I added call for target numbers to all sheet rolls and fixed
a small copy-paste error.