Adjustment of the base Characteristics rolls
Currently the base characteristics rolls were adding the value of the advance in that characteristic, but it wasn't reflected when using the skill rolls. This made this change necessary in order to align the sheet so both rolls have the correct target value.

Removed the advance from the Influence characteristic since it can't be upgraded this way like all other characteristics and this drop down menu is pointless.