Added mounts to the german char-sheet
Now you can enter 2 mounts according to the Rohan supplement and roll
for a charge as an attack. The riding modifier is automatically
substracted from the TN. The reduced encumbrance is not calculated
automatically as it applies only on armour and only when mounted - so
you have to do that manually via the inventory.