skills dice roll fix
I was playing with this sheet and came across with some bugs related to skill roll section. I saw Unarmed skill using different skill's value changed that to original as well as throwing skill. So they are now using their own values for rolling. Athleticism and barter skills have bugs too. When rolling for them in the chat we saw them as wrong name. Like Barter skill was the name of smallguns how evet their values are correct, so I only changed names.