GURPS (Style 1) - Further Updates
1) Added scripts to automatically fill point summeries.
2) Created function to handle summing up repeating section fields
3) Changed or removed default values of repeated sections to prevent bad calculations.
Note: For some reason when you create a new repeating item the "on change" event does not trigger, and fields that are filled in with a default value are read as undefined. For this reason I removed the default value for repeating sections. I also made the sumRepeating function assume a value of 0 for any falsey value.