Add Workerscript and remove roll2 and roll3
created a Workerscript to enable autocalculation of Damage Bonus and
Build even for people without a Pro DM.
Severely reduced logic around attribute and skill rolls but need to
improve the rolltemplates further. Got rid of Roll2 and Roll3 as only
one roll is needed for comparison.
Added derived value for dodge and own Language