Partial DFA.html revert
The change broke the roll buttons, but fixed the Fate Point calculations.

Rolling back the changes to the approaches.

Changed Mantle approach to fixed text because it never saved what was inside the text box. Also changed it to roll specifically for the mantle instead of Intellect as a fix from the first version.