VtM 20th AE i18n formatting test
I tried to create an i18n formatted character sheet to translate it in italian myself and permit other translations.
It should need a test to fix possible bugs due to my lacks of coding knowledge. 
I also created manually a .json translation file with the corrisponden keys used in this html.