Repeating Background Image for Talent and Spell Pages
Currently, when the Talent and Spell pages contain a large number of entries (Most Level 3+ Characters will have enough Talents/Spells to cause this issue), the background doesn't repeat and it simply goes to plain white. Which not only looks bad, but also makes entries that are on the edge of the background hard to read.

My proposed change is to simply have the backgrounds on those two pages repeat when they end, making the sheets look nicer and more uniform, and removing any legibility issues the current sheet may cause.