Update Zweihander.html
Enabled Basic, Intermediate and Advanced list buttons, selecting them would not save in roll20.  Corrected the Basic and Advanced point cost on the right hand side of the form, set them to 100rp and 300rp.  Changed 'Talent' heading to 'Trapping' heading in the second column of the Talents/Trappings section.  Set all the 'small attribute boxes' to 0 and used them to auto calculate racial attribute bonuses and penalties, you can now enter for example 1 or -1 in the small box and it will auto-calculate the new Attribute Bonus. Prior to this changing the attribute bonus manually to reflect racial modifications would not save.  Adding text specifying its use for Racial Mods and limiting the numbers entered from 2 to -2 would be helpful.