Delta Green RPG
New Character sheet for DG based on the rules provided in Need to Know and the Agent's Handbook by ArcDream Publishing.
Integrated roll template for the SAN check and all the attacks, skill and stat tests.
Support of the projection on bonds using the blue textbox next to the bond score.
Auto computation of the new Breaking Point with update every time the power of the character changes
Support of the automatic roll for special action like healing, resucitating, falling, catching fire, pinning, and disarm not supported in this version of the sheet.

These special rolls might be obtained rolling the relative skill and interpreting the output result.

Special thanks to AquaAlex in helping me creating the initial version