Add files via upload
Main Kata entry was missing the effects box.
Added sheet workers that set the selected approach as a output when you roll a skill.
Set the emotes for rolling a skill to be dumped from the translation file into variables, allowing them to be translated into other languages.
Sheet worker work required renaming variables in repeating sections. This is a breaking change, and one that disallows sheetworkers as a coversion method.
Added a note telling you how many dice you can keep on skill rolls and weapon attacks.