Beta Updates V3.0
Moved this sheet to its own branch so I could issue a pull request.
Changed Resilience to Endurance.
That's a derivative stat on PCs, so the new variable change will be invisible there.
Changed sheetworkers based on resilience variables to look at endurance instead.
Added a sheetworker to convert NPC_Resilience, which isn't a derivative stat. This will run once, the first time an NPC is opened.