[Infinity 2d20] Checkboxes to radio buttons; other minor additions
Changed the checkboxes in each stress track to radio buttons. This will allow the tying of token bars to each stress track.
Changed "Trauma" to "Metanoia" to reflect the change made when the RPG went to print.
Added a "Harm Effects" section and moved the "Traits" section to improve page flow.
Removed the "sheet-check-on" class: no longer used in sheet.