Update approved.yaml
See this thread for background (and also where Jakob suggest I edit the file and include comments to raise the issue):
https://app.roll20.net/forum/post/5983626/incorrect-categorization-fuzion-slash-cyberpunk-slash-hero-system/?pageforid=5983626

I am scared of being banned for editing this file, but this is an important issue, so I hope you'll be understanding. It looks to me like there is a difference in opinion behind the scenes because the groupings for these game systems keep being changed and then reverted. I don't know who is making the changes, so I can't contact them.
Anyway, the problem:
To summarise that thread: Hero System and Cyperpunk 2020 should definitely be separate groups.
Fusion (and bubblegum crisis) could be under CP2020, or be a separate group of its own. It should probably the latter (there are other games using the fuzion system that just havent made it on to roll20 yet).