Sentinel Comics RPG Sheet
Finally the RPG that is being Kickstarted this Summer 2018 has a bare bones sheet that can make play on Roll20 possible. Only the Starter Kit is currently available to the public so all elements of this sheet represent the game as it stands today. I used the information from the Starter kit here on this sheet.