Added 2 Features and Fixed 1 bug
Fixed the value for checkboxes was previously 0, is now 1 which allows the boxes to be unchecked.

I added a sheet worker which unchecks all the checkboxes so that it does not have to be done manually at the end of a session if desired.

Corrected the name of the Initiative (tactics) to be like the other fields on the sheet. It was backwards and listed as Tactics (initiative)