Fixed bugs - linked attributes and attributes improperly labled and not saving.
Fixed a bug where in repeating sections two different values were saving to the same atribute.
- This occured 3 times

Fixed 1 misnamed attribute which was pulling an incorrect value from somewhere else on the sheet.1

Fixed 5 attributes which were not properly tagged as attributs and not saving their values.

This affected ranged weapons to hit values, a fighting weapon to hit value, a field for where clothing is being worn,  strength and power values in  ship weapons, utilities, fixed components and internal components and vehicle weapon accuracy and to hit.