Item modifiers skills minor fix
Previously item modifiers for animal handling and sleight of hand did not properly apply skill bonuses. The expected string to trigger the changed used spaces whereas determining the value of the bonus was looking for the name of the skill to have underscores. This meant that it was impossible to use the item modifiers for this field.

This change will require the underscores between words for these skills so that the correct bonus can be applied.