Update 5th Edition OGL by Roll20.html
Hey Steve, your previous fix will work for the Animal Handling scenario, but it will not work for Sleight of Hand. Using a string for the replace method will only replace the first instance of the string that is found, not all of them. Changing the search string to a regex with the global tag will replace all instances of the string though.