Added item modifier for ability checks to 5e OGL
A new item modifier has been added to update all of the skills on the character sheet as well as the initiative for the character. Use the modifier, Ability Checks, to trigger this effect. This change allows for the utilization of items that modify all ability checks such as the Stone of Good Luck. However, this change does not impact the base ability checks, such as Strength checks.