Update WOIN.HTML to fix two typos, add new fields
This change would be to Fix the typo that when you roll Logic, it states you roll Agility. It will now say you roll Logic.
Reputation was written repuation everywhere. Made it Reputation again.
The sheet needed a field to add additional damage to a weapon roll, for example, when you want to roll 2d6+2. There was no way to add the +2. I added a field called Damage Add. for that.
The sheet needed a field to record current health and max health. Left the Health field as is and added a max Health field.