Reapply lost commits by oukag (#3786)
* Added item modifier for ability checks to 5e OGL

A new item modifier has been added to update all of the skills on the character sheet as well as the initiative for the character. Use the modifier, Ability Checks, to trigger this effect. This change allows for the utilization of items that modify all ability checks such as the Stone of Good Luck. However, this change does not impact the base ability checks, such as Strength checks.

* 5e OGL fixed spellcasting lvls of non-full casters

Half casters (Paladins and Rangers) and third casters (Arcane Tricksters and Eldritch Knights) will now calculate the correct number of spellslots. Previously the spell slots were rounding up to the next level of spellcasting when it should instead be rounded down.

* Spell slot calculation fix for multiclassing

Single class spellcasters round up in calculating the overall spellcasting level.
Multiple spellcasting classes round down when calculating the overall spellcasting level as mentioned on PHB pg 164.