Complete Redesign
This is a complete redesign and overhawl of the Godbound sheet to resolve several issues. There was a desire from the community for me to publish my work, as shown in this forum post: https://app.roll20.net/forum/post/6269089/new-godbound-sheet/?pageforid=6514817#post-6514817. I also received the original author's blessing to publish this new sheet and replace his work.