Transition Descript (notice)
There is a word difference between actual English and Korean.
Because it translated based on [character sheet] published in Korea.
I think it would be more convenient to use the word in [inSANe Rule Book] that was officially published in Korea for Korean people to use.

ex:
Auto Skill List 자동 스킬 목록 → Specialty List 특기 리스트
Skill List 스킬 목록 → Ability List 어빌리티 리스트