Replaced all instances of the wprd Fear or fear with the word Frifht. In GURPS the proper term is Fright or a Fright Check, not fear. Also modified the Fright check macro to replace the word Will wuth the word Fright so it states it is making the roll vs Fright and not vs Will.
Replaced all instances of the wprd Fear or fear with the word Frifht. In GURPS the proper term is Fright or a Fright Check, not fear. Also modified the Fright check macro to replace the word Will wuth the word Fright so it states it is making the roll vs Fright and not vs Will.