Wrath & Glory v2 release
Updates to HTML, CSS, readme, and json files.

Note this is a significant update to all files.  Version 1 of the character sheet was based on 28 pages of rules from a quickstart release.  This release is based on the core rulebook (452 pages).  The changes resulted in significant addition of content (nearly doubled in size) and a reorganization of existing content,   No existing attributes were changed or deleted; this was anticipated when the character sheet was created based on the quickstart rules.