Importing Buring's work
Thanks to @Burning!
Added Dodge and Parry full values to Defense section.
Powers, Advantages and Equipment now have a ROLL function. Clicking the Roll button will print the text of your power to the chat log as an emote, whisper to GM, whisper to self or public, as selected via the "Roll target - Effects" drop down in the Import and Options tab.