Update L5RFFGSheetCSS.css
Fixed recent break to Composure and Endurance Calculation, while maintaining the attempt to correct the new books swap of the stat calculations...

Updated the wording on Adversities and Anxieties to reflect the official release character sheet.

Changed "status" to "standing" in the section on other characters to reflect the wording on the official release character sheet.

Although not every stance changed mechanically, changed the tooltips on every stance to reflect the wording on the new, official character sheet, post-release.

Added sections for ninjutsu and maho technique to the PC sheet. NPCs will follow.