Correct the alignment of the Skills tab
It seems that the use of DIVs over TABLEs on this table defeated the
usability in game. This uses TABLEs to present the TABULAR list of
Skills with Mastery, custom mods, the calculated Sum, and Roll
button.

Some inline style sheet work was cleaned up and put in the CSS.