Enhancements to the AS&SH Character Sheet (
Add attack and damage dice rollers to the Melee and Missile weapons section.  The result should show the target AC hit and damage caused.
Change Saving Throws from a single dice roller to a dice roller per saving throw type. 
Add a dice roller into the spell section. Not all spells need it, but it'd be nice if a roller could return the spell name and a damage or duration roll that is input.
Change the "Gems and Jewelry" section of the Coinage to a repeating section with fields for Type, Value, Quantity.