Added "Things from the Flood" tab (#4817)
This sheet now works for both "Tales from the Loop" and "Things from the
Flood". Could you please update the sheet name in the dropdown (perhaps
to "TFTL/TFTF") to indicate the change?

I also did some minor cleanup on the sheet, changing some french words
to english and updating the translation files.