UESTRPG sheet redone (#5017)
* Added UESTRPG based on the 5e Sheet

This is a customised version of the 5e OGL sheet for use in the Unofficial Elder Scrolls Tabletop Role-Playing Game conversion. Known issues: Agility saves for NPCs don't display properly, minor formatting issue with temp HP

* Added spell costs, fixed NPC AGI save bug

* Create UESTRPG core page preview.png

* Added preview, user ID, clarified file names

* Update sheet.json

* Renamed folder

* Corrected a bug that caused some classes to be assigned wrong hit die

* Update UESTRPG by EnricoDandolo HTML.html

* Reverted some changes that broke compendium import of monster stablocks

* Remade the sheet from scratch after encountering a number of issues

The previous version was beset by issues and marred by flaws. This time, I sought to stay as close as possible to the OGL sheet's code in order.

* Remade the sheet from scratch after encountering a number of issues

The previous version was beset by issues and marred by flaws. This time, I sought to stay as close as possible to the OGL sheet's code in order.

* Let's try this again ...