changes to Luck Attribute
Luck is a special attribute that functions a little differently from the rest. I've taken away a few fields that are not associated with Luck, modified the roll template to account for these changes, and added a dropdown for luck (attribute) saves. New column width style associated with luck added to css as well.