Tremulus RPG (#5258)
* Tremulus RPG

Added a sheet for the Tremulus RPG.  This is another one that is powered by the apocalypse

* Bug fix and translation start

Corrected a bug where many of the selected moves were linked.  They should now all be individuals.

Also started down the path of adding translation support.  Tags were added for the basic character details, as well as the left column of the playset.

The right column (moves) and all of the rolling macros still need marked up.

* Sheet translation

All the text on the sheet is marked up for translation.

None of the sheet macros are done though, need to look into how those would work.

* Translation done

All the translaction text should be in place.  All sheet values as well as visible macro text is accounted for.