[Band of Blades] Add Band of Blades Official v1.0
This includes complete sheets conforming to the release version of the game, including legion roles and Broken/Chosen. This sheet was created in collaboration with Evil Hat Games, so I have the necessary permission of all the copyright holders.

The new sheet is *not* compatible with the existing sheet for Band of Blades. I would suggest that the old sheet be deprecated in favor of this one.