Bug fixes for sheets: Delta Green, Monsterhearts-2, and Tremulus (#5291)
* Tremulus RPG

Added a sheet for the Tremulus RPG.  This is another one that is powered by the apocalypse

* Bug fix and translation start

Corrected a bug where many of the selected moves were linked.  They should now all be individuals.

Also started down the path of adding translation support.  Tags were added for the basic character details, as well as the left column of the playset.

The right column (moves) and all of the rolling macros still need marked up.

* Sheet translation

All the text on the sheet is marked up for translation.

None of the sheet macros are done though, need to look into how those would work.

* Translation done

All the translaction text should be in place.  All sheet values as well as visible macro text is accounted for.

* Delta green bug fix

Adding a fix for one of the bugs posted on the Roll20 Char Sheet git hub

* Fixed move attributes

The attr_ prefix on the checkboxes for moves had a type, using a hypen instead of an underscore.  This prevented the values from being saved.

* Translating Monsterhearts

Adding a translation file for the monsterhearts 2 sheet.  Updated lots of the HTML to have the needed tags.

Missing magic moves (witch hexes and infernal bargains), darkest self text, and all the stuff for experience and advancements.

* Sheet JSON rename

Had a file named wrong.

* Finished all sheet translactions

THe macros still need some work, both in the prompts as well as the displayed text.  Since images are used for some of the roll stuff that might be much trickier.

* Translation done

Corrected the sheet macros to use translatable text.  That should be the end of the translation work for Monsterhearts 2