[Palladium Fantasy 1E] sheet (#5400)
* Update palladium_fantasy_1e.html

Created a new tab 2.  Moved some content from tab 1 to tab 2 in order to reduce clutter on tab 1.  Restructured tab 1 and tab 5.

* Update palladium_fantasy_1e.css

Created a new tab 2, restructured tab 1, added borders to areas of tab 1 and 5, and changed roll button decorations.

* Added PalStP1.png

This is the new image for tab 1 of the character sheet.

* Delete PalStP1.png

* Add files via upload

* Delete palladium_fantasy_1e.png

* Update sheet.json

Changed preview image to reflect changes in sheet.

* Update Readme.md

Made a few edits to reflect changes in the sheet.