Low fantasy gaming updates (#5609)
* Bugfix:  Initiative adds bonus to target DC

Rolling initiative with the deluxe rules now proper adds Dex+Misc Bonus to target DC rather than as a modifier to the roll.

* Bugfix:  Luck now adds bonuses to target DC

Luck rolls now add bonuses to the target DC instead of as a modifier to the roll.

* Bugfix:  Luck rolls now check for minimum luck.

Luck rolls now check that your Luck rolls are using the proper minimum of 1/2 your level.

* Feature:  GS and TF are automatically calculated

Great Success and Terrible Failure DC values no longer need to be entered manually.  They are automatically calculated from your attribute scores.

* Feature:  Add XP and XP goal fields

In preparation for the optional XP rules provided in the LFG Companion book, XP and XP goal fields have been added.

* Bugfix:  Level field now has default value of 1

Level field didn't have a default value, and this caused a syntax error with luck saves if a level is never added.

* Bugfix:  Capped "Terrible Failure" score at 20.

* Add maximum target DC for attribute checks

In Low Fantasy Gaming, attribute checks are a d20, roll under, system.  No matter how high your attribute, a natural roll of 20 is a failure.  Natural 20's were not presenting as failures in the case of an attribute score higher than 19.