Kagegami (#5593)
* Kagegami High Character Sheet version 1.0

This is the first submission of a character sheet for the game Kagegami High.

* Corrected Spelling Error

Corrected spelling error in titile of character sheet.

Additionally, this character sheet was approved four months ago, but remains unviewable in Roll20.  People on the forum suggest the issue is that it was not added to approved.yaml.  Thoughts?

* Errors Corrected

Corrected file name for PR and changes JSON file to align names with file names.