[Palladium Fantasy 1E] update (#5641)
* Delete sheet_tab1.png

This will be replaced with an updated image

* Add a new sheet_tab1.png

This is an updated image with the public/whisper switch.

* Update palladium_fantasy_1e.html

Added a public/whisper roll toggle to the top of tab 1.  All macros in the sheet were edited to check for the whisper.

* Update palladium_fantasy_1e.css

Added the whisper toggle at the top of tab 1.  All changes are for that toggle and its labeling