Update palladium_fantasy_1e.html
Moved temporary bonuses out of the repeating sections for melee and ranged attacks.  Made the roll buttons in those sections larger to facilitate a full label.  Hid the remaining entry fields with a checkbox at the end of each row to make them visible as a second line.  Changed the "notes" input field to a textarea so that it would allow line breaks, since the description part of the roll template can accept those.