Update [Palladium Fantasy 1E] (#5866)
* Update palladium_fantasy_1e.html

Updated the release notes section of the sheet.  Clearly marked the beginning of each tab of the sheet in the code to help make things easier for anyone following after..

* Update palladium_fantasy_1e.css

Attempted to create some notes for different sections of the code to make things easier for anyone following after me.  Updated the look of repeating section control buttons to fit in better with the rest of the sheet.

* Delete sheet_tab1.png

* Delete sheet_tab2.png

* Delete sheet_tab3.png

* Delete sheet_tab4.png

* Delete sheet_tab5.png

* Delete sheet_tab6.png

* Add files via upload

Updated images of the character sheet tabs.